Unit tst structure update
