+++ src/raspberrypi/test/test_setup.cpp
@@ -8,7 +8,10 @@
 //---------------------------------------------------------------------------
 
 #include <gtest/gtest.h>
+
 #include "spdlog/spdlog.h"
+#include "controllers/controller_manager.h"
+#include "devices/device_factory.h"
 
 class Environment : public ::testing::Environment
 {
@@ -20,6 +23,10 @@ public:
 	// Turn off logging
 	void SetUp() override { spdlog::set_level(spdlog::level::off); }
 };
+
+DeviceFactory& device_factory = DeviceFactory::instance();
+
+ControllerManager& controller_manager = ControllerManager::instance();
 
 int main(int, char*[])
 {
